Updated README, changed filename
